Fix(Web): use only validated safe address
