+++ apps/web/src/hooks/useSafeAddress.ts
@@ -1,19 +1,8 @@
-import { useMemo } from 'react'
-import { useRouter } from 'next/router'
-import { parsePrefixedAddress } from '@/utils/addresses'
+import useSafeInfo from '@/hooks/useSafeInfo'
 
 const useSafeAddress = (): string => {
-  const router = useRouter()
-  const { safe = '' } = router.query
-  const fullAddress = Array.isArray(safe) ? safe[0] : safe
-
-  const checksummedAddress = useMemo(() => {
-    if (!fullAddress) return ''
-    const { address } = parsePrefixedAddress(fullAddress)
-    return address
-  }, [fullAddress])
-
-  return checksummedAddress
+  const { safeAddress } = useSafeInfo()
+  return safeAddress
 }
 
 export default useSafeAddress
